chore: buttonMessage breakdown into components
